feat: Update app.py
