+++ frontend/requirements.txt
@@ -1,0 +1,5 @@
+streamlit
+supabase
+st_login_form
+st-supabase-connection
+argon2
